Change all float to double.

+++ include/geometry/types.hpp
@@ -7,30 +7,30 @@ namespace PoIntInt {
 
 // Types needed for CUDA computation
 struct TriPacked {
-  float3 a;   // one vertex
-  float3 e1;  // b - a
-  float3 e2;  // c - a
-  float3 S;   // 0.5 * (e1 x e2)  (oriented area vector)
+  double3 a;   // one vertex
+  double3 e1;  // b - a
+  double3 e2;  // c - a
+  double3 S;   // 0.5 * (e1 x e2)  (oriented area vector)
 };
 
 // Oriented point cloud element (circular disk surfel)
-// Note: Struct size is 32 bytes (12+12+4+4), naturally aligned
+// Note: Struct size is 48 bytes (24+24+8+8), naturally aligned
 struct DiskPacked {
-  float3 c;   // center (12 bytes)
-  float3 n;   // unit normal (12 bytes)
-  float rho;  // radius (4 bytes)
-  float area; // area weight (π * rho^2 for disk) (4 bytes)
-  // Total: 32 bytes
+  double3 c;   // center (24 bytes)
+  double3 n;   // unit normal (24 bytes)
+  double rho;  // radius (8 bytes)
+  double area; // area weight (π * rho^2 for disk) (8 bytes)
+  // Total: 64 bytes
 };
 
 // Gaussian splat surfel (planar Gaussian footprint)
-// Note: Struct size is 32 bytes (12+12+4+4), naturally aligned
+// Note: Struct size is 48 bytes (24+24+8+8), naturally aligned
 struct GaussianPacked {
-  float3 c;   // center (12 bytes)
-  float3 n;   // unit normal (12 bytes)
-  float sigma; // standard deviation in tangent plane (4 bytes)
-  float w;     // area weight (4 bytes)
-  // Total: 32 bytes
+  double3 c;   // center (24 bytes)
+  double3 n;   // unit normal (24 bytes)
+  double sigma; // standard deviation in tangent plane (8 bytes)
+  double w;     // area weight (8 bytes)
+  // Total: 64 bytes
 };
 
 // Geometry type enum
